+++ processor/src/lib.rs
@@ -409,10 +409,14 @@ where
     /// expected to be either in the current `program` or in the host.
     #[inline(always)]
     fn execute_dyn_node(&mut self, program: &MastForest) -> Result<(), ExecutionError> {
-        self.start_dyn_node()?;
+        // TODO(plafer): Refactor. We need to call those before `start_dyn_node()`, since
+        // `start_dyn_node()` advances the clock.
+        let mem_addr = self.stack.get(0);
+        // The callee hash is stored in memory, and the address is specified on the top of the
+        // stack.
+        let callee_hash = self.read_mem_word(mem_addr)?;
 
-        // get target hash from the stack
-        let callee_hash = self.stack.get_word(0);
+        self.start_dyn_node(callee_hash)?;
 
         // if the callee is not in the program's MAST forest, try to find a MAST forest for it in
         // the host (corresponding to an external library loaded in the host); if none are