+++ processor/src/stack/aux_trace.rs
@@ -21,6 +21,8 @@ impl AuxTraceBuilder {
         rand_elements: &[E],
     ) -> Vec<Vec<E>> {
         let p1 = self.build_aux_column(main_trace, rand_elements);
+
+        debug_assert_eq!(*p1.last().unwrap(), E::ONE);
         vec![p1]
     }
 }